icu_segmenter

Module grapheme

Structs

GraphemeClusterBreakIterator

Implements the Iterator trait over the grapheme cluster boundaries of the given string.

GraphemeClusterSegmenter

Segments a string into grapheme clusters.

Type Aliases

GraphemeClusterBreakIteratorLatin1

Grapheme cluster break iterator for a Latin-1 (8-bit) string.

GraphemeClusterBreakIteratorPotentiallyIllFormedUtf8

Grapheme cluster break iterator for a potentially invalid UTF-8 string.

GraphemeClusterBreakIteratorUtf8

Grapheme cluster break iterator for an str (a UTF-8 string).

GraphemeClusterBreakIteratorUtf16

Grapheme cluster break iterator for a UTF-16 string.