icu_segmenter

Module line

LineSegmenter and its related iterators, borrowed types, and options.

Structs

LineBreakIterator

Implements the Iterator trait over the line break opportunities of the given string.

LineBreakOptions

Options to tailor line-breaking behavior.

LineSegmenter

Supports loading line break data, and creating line break iterators for different string encodings.

LineSegmenterBorrowed

Segments a string into lines (borrowed version).

ResolvedLineBreakOptions 🔒

Enums

LineBreakStrictness

An enum specifies the strictness of line-breaking rules. It can be passed as an argument when creating a line segmenter.

LineBreakWordOption

An enum specifies the line break opportunities between letters. It can be passed as an argument when creating a line segmenter.

StringBoundaryPosType 🔒

Constants

AI 🔒

AK 🔒

AL 🔒

AL_DOTTED_CIRCLE 🔒

AP 🔒

AS 🔒

B2 🔒

BA 🔒

BB 🔒

BK 🔒

CB 🔒

CJ 🔒

CL 🔒

CM 🔒

CP 🔒

CR 🔒

EB 🔒

EM 🔒

EX 🔒

GL 🔒

H2 🔒

H3 🔒

HL 🔒

HY 🔒

ID 🔒

ID_CN 🔒

IN 🔒

IS 🔒

JL 🔒

JT 🔒

JV 🔒

LF 🔒

NL 🔒

NS 🔒

NU 🔒

OP_EA 🔒

OP_OP30 🔒

PO 🔒

PO_EAW 🔒

PR 🔒

PR_EAW 🔒

QU 🔒

QU_PF 🔒

QU_PI 🔒

RI 🔒

SA 🔒

SP 🔒

SY 🔒

UNKNOWN 🔒

VF 🔒

VI 🔒

WJ 🔒

XX 🔒

ZW 🔒

ZWJ 🔒

Traits

LineBreakType

A trait allowing for LineBreakIterator to be generalized to multiple string iteration methods.

Functions

is_break_utf32_by_loose 🔒

line_handle_complex_language_utf8 🔒

line_handle_complex_language impl for UTF8 iterators