icu_segmenter

Module rule_segmenter

Structs

Latin1

RuleBreakType for Latin-1 strings

PotentiallyIllFormedUtf8

RuleBreakType for potentially ill-formed UTF-8 strings

RuleBreakIterator

Implements the Iterator trait over the segmenter boundaries of the given string.

Utf8

RuleBreakType for UTF-8 strings

Utf16

RuleBreakType for UTF-16 strings

Traits

RuleBreakType

A trait allowing for RuleBreakIterator to be generalized to multiple string encoding methods and granularity such as grapheme cluster, word, etc.

Functions

empty_handle_complex_language 🔒