icu_segmenter::rule_segmenter

Trait RuleBreakType

pub trait RuleBreakType<'l, 's> {
    type IterAttr: Iterator<Item = (usize, Self::CharType)> + Clone + Debug;
    type CharType: Copy + Into<u32> + Debug;

    // Required methods
    fn get_current_position_character_len(
        iter: &RuleBreakIterator<'l, 's, Self>,
    ) -> usize;
    fn handle_complex_language(
        iter: &mut RuleBreakIterator<'l, 's, Self>,
        left_codepoint: Self::CharType,
    ) -> Option<usize>;
}

A trait allowing for RuleBreakIterator to be generalized to multiple string encoding methods and granularity such as grapheme cluster, word, etc.

Required Associated Types

type IterAttr: Iterator<Item = (usize, Self::CharType)> + Clone + Debug

The iterator over characters.

type CharType: Copy + Into<u32> + Debug

The character type.

Required Methods

fn get_current_position_character_len( iter: &RuleBreakIterator<'l, 's, Self>, ) -> usize

fn handle_complex_language( iter: &mut RuleBreakIterator<'l, 's, Self>, left_codepoint: Self::CharType, ) -> Option<usize>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.