icu_segmenter

Struct WordSegmenter

pub struct WordSegmenter(/* private fields */);

Supports loading word break data, and creating word break iterators for different string encodings.

Most segmentation methods live on WordSegmenterBorrowed, which can be obtained via [WordSegmenter::new_auto()] (etc) or WordSegmenter::as_borrowed().

Word segmenter is currently compatible with Unicode Standard Annex #29 (Version 17.0.0).

Content Locale

You can optionally provide a content locale to the WordSegmenter constructor. If you have information on the language of the text being segmented, providing this hint can produce higher-quality results.

If you have a content locale, use WordBreakOptions and a constructor beginning with try_new. If you do not have a content locale use WordBreakInvariantOptions and a constructor beginning with new.

Examples

Segment a string:

use icu::segmenter::WordSegmenter;

let segmenter = WordSegmenter::new_auto(Default::default());

let breakpoints: Vec<usize> =
    segmenter.segment_str("Hello World").collect();
assert_eq!(&breakpoints, &[0, 5, 6, 11]);

Segment a Latin1 byte string with a content locale:

use icu::locale::langid;
use icu::segmenter::WordSegmenter;
use icu::segmenter::options::WordBreakOptions;

let mut options = WordBreakOptions::default();
let langid = &langid!("en");
options.content_locale = Some(langid);
let segmenter = WordSegmenter::try_new_auto(options).unwrap();

let breakpoints: Vec<usize> = segmenter
    .as_borrowed()
    .segment_latin1(b"Hello World")
    .collect();
assert_eq!(&breakpoints, &[0, 5, 6, 11]);

Successive boundaries can be used to retrieve the segments. In particular, the first boundary is always 0, and the last one is the length of the segmented text in code units.

use itertools::Itertools;
let text = "Mark’d ye his words?";
let segments: Vec<&str> = segmenter
    .segment_str(text)
    .tuple_windows()
    .map(|(i, j)| &text[i..j])
    .collect();
assert_eq!(
    &segments,
    &["Mark’d", " ", "ye", " ", "his", " ", "words", "?"]
);

Not all segments delimited by word boundaries are words; some are interword segments such as spaces and punctuation. The WordBreakIterator::word_type() of a boundary can be used to classify the preceding segment; WordBreakIterator::iter_with_word_type() associates each boundary with its status.

let words: Vec<&str> = segmenter
    .segment_str(text)
    .iter_with_word_type()
    .tuple_windows()
    .filter(|(_, (_, segment_type))| segment_type.is_word_like())
    .map(|((i, _), (j, _))| &text[i..j])
    .collect();
assert_eq!(&words, &["Mark’d", "ye", "his", "words"]);

Implementations

impl WordSegmenter

pub fn new_dictionary( options: WordBreakInvariantOptions, ) -> WordSegmenterBorrowed<'static>

Construct a WordSegmenter with an invariant locale and compiled dictionary data for complex scripts (Chinese, Japanese, Khmer, Lao, Myanmar, and Thai).

The dictionary model uses a list of words to determine appropriate breakpoints. It is faster than the LSTM model but requires more data.

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

Examples

Behavior with complex scripts:

use icu::segmenter::{WordSegmenter, options::WordBreakInvariantOptions};

let th_str = "ทุกสองสัปดาห์";
let ja_str = "こんにちは世界";

let segmenter =
    WordSegmenter::new_dictionary(WordBreakInvariantOptions::default());

let th_bps = segmenter.segment_str(th_str).collect::<Vec<_>>();
let ja_bps = segmenter.segment_str(ja_str).collect::<Vec<_>>();

assert_eq!(th_bps, [0, 9, 18, 39]);
assert_eq!(ja_bps, [0, 15, 21]);

pub fn try_new_dictionary( options: WordBreakOptions<'_>, ) -> Result<Self, DataError>

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

pub fn try_new_dictionary_unstable<D>( provider: &D, options: WordBreakOptions<'_>, ) -> Result<Self, DataError>
where D: DataProvider<SegmenterBreakWordV1> + DataProvider<SegmenterBreakWordOverrideV1> + DataProvider<SegmenterDictionaryAutoV1> + DataProvider<SegmenterDictionaryExtendedV1> + DataProvider<SegmenterBreakGraphemeClusterV1> + ?Sized,

A version of Self::new_dictionary that uses custom data provided by a DataProvider.

📚 Help choosing a constructor

⚠️ The bounds on provider may change over time, including in SemVer minor releases.

pub const fn new_for_non_complex_scripts( _options: WordBreakInvariantOptions, ) -> WordSegmenterBorrowed<'static>

Construct a WordSegmenter with an invariant locale and no support for scripts requiring complex context dependent word breaks (Chinese, Japanese, Khmer, Lao, Myanmar, and Thai).

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

pub fn try_new_for_non_complex_scripts( options: WordBreakOptions<'_>, ) -> Result<Self, DataError>

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

pub fn try_new_for_non_complex_scripts_unstable<D>( provider: &D, options: WordBreakOptions<'_>, ) -> Result<Self, DataError>
where D: DataProvider<SegmenterBreakWordV1> + DataProvider<SegmenterBreakWordOverrideV1> + DataProvider<SegmenterBreakGraphemeClusterV1> + ?Sized,

A version of Self::new_for_non_complex_scripts that uses custom data provided by a DataProvider.

📚 Help choosing a constructor

⚠️ The bounds on provider may change over time, including in SemVer minor releases.

pub fn load_dictionary_unstable<D>( &mut self, provider: &D, ) -> Result<(), DataError>
where D: DataProvider<SegmenterDictionaryAutoV1> + DataProvider<SegmenterDictionaryExtendedV1> + ?Sized,

Loads dictionary data for a WordSegmenter constructed with WordSegmenter::new_for_non_complex_scripts.

pub fn as_borrowed(&self) -> WordSegmenterBorrowed<'_>

Constructs a borrowed version of this type for more efficient querying.

Most useful methods for segmentation are on this type.