icu_segmenter

Module word

WordSegmenter and its related iterators, borrowed types, and options.

Re-exports

pub use inner::WordType;

Modules

inner 🔒

Hide ULE type

Structs

WordBreakInvariantOptions

Locale-independent options to tailor word breaking behavior

WordBreakIterator

Implements the Iterator trait over the word boundaries of the given string.

WordBreakIteratorWithWordType

Word break iterator that also returns the word type

WordBreakOptions

Options to tailor word breaking behavior.

WordSegmenter

Supports loading word break data, and creating word break iterators for different string encodings.

WordSegmenterBorrowed

Segments a string into words (borrowed version).

Traits

WordBreakType

A trait allowing for WordBreakIterator to be generalized to multiple string iteration methods.

Functions

handle_complex_language_utf8 🔒

handle_complex_language impl for UTF8 iterators