Crate icu_segmenter

A segmenter implementation for the following rules.

Line breaker that is compatible with Unicode Standard Annex #14 and CSS properties.

Grapheme cluster breaker, word breaker, and sentence breaker that are compatible with Unicode Standard Annex #29.

Examples

Line Break

Segment a string with default options:

 use icu_segmenter::LineBreakSegmenter;

 let provider = icu_provider::inv::InvariantDataProvider;
 let segmenter = LineBreakSegmenter::try_new(&provider)
     .expect("Data exists");

 let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
 assert_eq!(&breakpoints, &[6, 11]);

Segment a string with CSS option overrides:

use icu_segmenter::{LineBreakSegmenter, LineBreakOptions, LineBreakRule, WordBreakRule};

let mut options = LineBreakOptions::default();
options.line_break_rule = LineBreakRule::Strict;
options.word_break_rule = WordBreakRule::BreakAll;
options.ja_zh = false;
let provider = icu_provider::inv::InvariantDataProvider;
let segmenter = LineBreakSegmenter::try_new_with_options(&provider, options)
    .expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
assert_eq!(&breakpoints, &[1, 2, 3, 4, 6, 7, 8, 9, 10, 11]);

Segment a Latin1 byte string:

use icu_segmenter::LineBreakSegmenter;

let provider = icu_provider::inv::InvariantDataProvider;
let segmenter = LineBreakSegmenter::try_new(&provider)
    .expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_latin1(b"Hello World").collect();
assert_eq!(&breakpoints, &[6, 11]);

Grapheme Cluster Break

Segment a string:

 use icu_segmenter::GraphemeClusterBreakSegmenter;
 let segmenter = GraphemeClusterBreakSegmenter::try_new()
     .expect("Data exists");

 let breakpoints: Vec<usize> = segmenter.segment_str("Hello 🗺").collect();
 // World Map (U+1F5FA) is encoded in four bytes in UTF-8.
 assert_eq!(&breakpoints, &[0, 1, 2, 3, 4, 5, 6, 10]);

Segment a Latin1 byte string:

use icu_segmenter::GraphemeClusterBreakSegmenter;
let segmenter = GraphemeClusterBreakSegmenter::try_new()
    .expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_latin1(b"Hello World").collect();
assert_eq!(&breakpoints, &[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]);

Word Break

Segment a string:

 use icu_segmenter::WordBreakSegmenter;
 let segmenter = WordBreakSegmenter::try_new()
     .expect("Data exists");

 let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
 assert_eq!(&breakpoints, &[0, 5, 6, 11]);

Segment a Latin1 byte string:

use icu_segmenter::WordBreakSegmenter;
let segmenter = WordBreakSegmenter::try_new()
    .expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_latin1(b"Hello World").collect();
assert_eq!(&breakpoints, &[0, 5, 6, 11]);

Sentence Break

Segment a string:

 use icu_segmenter::SentenceBreakSegmenter;
 let segmenter = SentenceBreakSegmenter::try_new()
     .expect("Data exists");

 let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
 assert_eq!(&breakpoints, &[0, 11]);

Segment a Latin1 byte string:

use icu_segmenter::SentenceBreakSegmenter;
let segmenter = SentenceBreakSegmenter::try_new()
    .expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_latin1(b"Hello World").collect();
assert_eq!(&breakpoints, &[0, 11]);

Modules

provider

Data provider struct definitions for this ICU4X component.

Structs

GraphemeClusterBreakSegmenter

Supports loading grapheme cluster break data, and creating grapheme cluster break iterators for different string encodings. Please see the module-level documentation for its usages.

Latin1Char

LineBreakIterator

Implements the Iterator trait over the line break opportunities of the given string. Please see the module-level documentation for its usages.

LineBreakOptions

Options to tailor line breaking behavior, such as for CSS.

LineBreakSegmenter

Supports loading line break data, and creating line break iterators for different string encodings. Please see the module-level documentation for its usages.

SentenceBreakSegmenter

Supports loading sentence break data, and creating sentence break iterators for different string encodings. Please see the module-level documentation for its usages.

Utf16Char

WordBreakSegmenter

Supports loading word break data, and creating word break iterators for different string encodings. Please see the module-level documentation for its usages.

Enums

LineBreakRule

An enum specifies the strictness of line-breaking rules. It can be passed as an argument when creating a line breaker.

WordBreakRule

An enum specifies the line break opportunities between letters. It can be passed as an argument when creating a line breaker.

Constants

AI

AL

B2

B2_SP

BA

BB

BK

BREAK_PROPERTIES_0

BREAK_PROPERTIES_1

BREAK_PROPERTIES_2

BREAK_PROPERTIES_3

BREAK_PROPERTIES_4

BREAK_PROPERTIES_5

BREAK_PROPERTIES_6

BREAK_PROPERTIES_7

BREAK_PROPERTIES_8

BREAK_PROPERTIES_9

BREAK_PROPERTIES_10

BREAK_PROPERTIES_11

BREAK_PROPERTIES_12

BREAK_PROPERTIES_13

BREAK_PROPERTIES_14

BREAK_PROPERTIES_15

BREAK_PROPERTIES_16

BREAK_PROPERTIES_17

BREAK_PROPERTIES_18

BREAK_PROPERTIES_19

BREAK_PROPERTIES_20

BREAK_PROPERTIES_21

BREAK_PROPERTIES_22

BREAK_PROPERTIES_23

BREAK_PROPERTIES_24

BREAK_PROPERTIES_25

BREAK_PROPERTIES_26

BREAK_PROPERTIES_27

BREAK_PROPERTIES_28

BREAK_PROPERTIES_29

BREAK_PROPERTIES_30

BREAK_PROPERTIES_31

BREAK_PROPERTIES_32

BREAK_PROPERTIES_33

BREAK_PROPERTIES_34

BREAK_PROPERTIES_35

BREAK_PROPERTIES_36

BREAK_PROPERTIES_37

BREAK_PROPERTIES_38

BREAK_PROPERTIES_39

BREAK_PROPERTIES_40

BREAK_PROPERTIES_41

BREAK_PROPERTIES_42

BREAK_PROPERTIES_43

BREAK_PROPERTIES_44

BREAK_PROPERTIES_45

BREAK_PROPERTIES_46

BREAK_PROPERTIES_47

BREAK_PROPERTIES_48

BREAK_PROPERTIES_49

BREAK_PROPERTIES_50

BREAK_PROPERTIES_51

BREAK_PROPERTIES_52

BREAK_PROPERTIES_53

BREAK_PROPERTIES_54

BREAK_PROPERTIES_55

BREAK_PROPERTIES_56

BREAK_PROPERTIES_57

BREAK_PROPERTIES_58

BREAK_PROPERTIES_59

BREAK_PROPERTIES_60

BREAK_PROPERTIES_61

BREAK_PROPERTIES_62

BREAK_PROPERTIES_63

BREAK_PROPERTIES_64

BREAK_PROPERTIES_65

BREAK_PROPERTIES_66

BREAK_PROPERTIES_67

BREAK_PROPERTIES_68

BREAK_PROPERTIES_69

BREAK_PROPERTIES_70

BREAK_PROPERTIES_71

BREAK_PROPERTIES_72

BREAK_PROPERTIES_73

BREAK_PROPERTIES_74

BREAK_PROPERTIES_75

BREAK_PROPERTIES_76

BREAK_PROPERTIES_77

BREAK_PROPERTIES_78

BREAK_PROPERTIES_79

BREAK_PROPERTIES_80

BREAK_PROPERTIES_81

BREAK_PROPERTIES_82

BREAK_PROPERTIES_83

BREAK_PROPERTIES_84

BREAK_PROPERTIES_85

BREAK_PROPERTIES_86

BREAK_PROPERTIES_87

BREAK_PROPERTIES_88

BREAK_PROPERTIES_89

BREAK_PROPERTIES_90

BREAK_PROPERTIES_91

BREAK_PROPERTIES_92

BREAK_PROPERTIES_93

BREAK_PROPERTIES_94

BREAK_PROPERTIES_95

BREAK_PROPERTIES_96

BREAK_PROPERTIES_97

BREAK_PROPERTIES_98

BREAK_PROPERTIES_99

BREAK_PROPERTIES_100

BREAK_PROPERTIES_101

BREAK_PROPERTIES_102

BREAK_PROPERTIES_103

BREAK_PROPERTIES_104

BREAK_PROPERTIES_105

BREAK_PROPERTIES_106

BREAK_PROPERTIES_107

BREAK_PROPERTIES_108

BREAK_PROPERTIES_109

BREAK_PROPERTIES_110

BREAK_PROPERTIES_111

BREAK_PROPERTIES_112

BREAK_PROPERTIES_113

BREAK_PROPERTIES_114

BREAK_PROPERTIES_115

BREAK_PROPERTIES_116

BREAK_PROPERTIES_117

BREAK_PROPERTIES_118

BREAK_PROPERTIES_119

BREAK_PROPERTIES_120

BREAK_PROPERTIES_121

BREAK_PROPERTIES_122

BREAK_PROPERTIES_123

BREAK_PROPERTIES_124

BREAK_PROPERTIES_125

BREAK_PROPERTIES_126

BREAK_PROPERTIES_127

BREAK_PROPERTIES_FILL_BY_0

BREAK_PROPERTIES_FILL_BY_ID

BREAK_PROPERTIES_FILL_BY_XX

BREAK_RULE

BREAK_STATE_MACHINE_TABLE

CB

CJ

CL

CL_CP_SP

CM

CP

CR

EB

EM

EOT

EX

GL

H2

H3

HL

HL_HY

HY

ID

ID_CN

IN

INTERMEDIATE_MATCH_RULE

IS

JL

JT

JV

KEEP_RULE

LAST_CODEPOINT_PROPERTY

LB25_HY

LB25_NU_CL

LB25_NU_CP

LB25_NU_IS

LB25_NU_SY

LB25_OP

LF

NL

NOT_MATCH_RULE

NS

NU

OP_EA

OP_OP30

OP_SP

PO

PO_EAW

PR

PROPERTY_COUNT

PROPERTY_TABLE

PROP_COMPLEX

PROP_EOT

PROP_SOT

PR_EAW

QU

QU_SP

RI

RI_RI

SA

SG

SOT

SP

SY

UNKNOWN

WJ

XX

ZW

ZWJ

Traits

LineBreakType

A trait allowing for LineBreakIterator to be generalized to multiple string iteration methods.

Type Definitions

GraphemeClusterBreakIterator

Grapheme cluster break iterator for an str (a UTF-8 string).

GraphemeClusterBreakIteratorLatin1

Grapheme cluster break iterator for a Latin-1 (8-bit) string.

GraphemeClusterBreakIteratorUtf16

Grapheme cluster break iterator for a UTF-16 string.

SentenceBreakIterator

Sentence break iterator for an str (a UTF-8 string).

SentenceBreakIteratorLatin1

Sentence break iterator for a Latin-1 (8-bit) string.

SentenceBreakIteratorUtf16

Sentence break iterator for a UTF-16 string.

WordBreakIterator

Word break iterator for an str (a UTF-8 string).

WordBreakIteratorLatin1

Word break iterator for a Latin-1 (8-bit) string.

WordBreakIteratorUtf16

Word break iterator for a UTF-16 string.