Crate icu_segmenter

A segmenter implementation for the following rules.

Line breaker that is compatible with Unicode Standard Annex #14 and CSS properties.

Word breaker that is compatible with Unicode Standard Annex #29.

Examples

Line Break

Segment a string with default options:

 use icu_segmenter::LineBreakSegmenter;

 let segmenter = LineBreakSegmenter::try_new().expect("Data exists");

 let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
 assert_eq!(&breakpoints, &[6, 11]);

Segment a string with CSS option overrides:

use icu_segmenter::{LineBreakSegmenter, LineBreakOptions, LineBreakRule, WordBreakRule};

let mut options = LineBreakOptions::default();
options.line_break_rule = LineBreakRule::Strict;
options.word_break_rule = WordBreakRule::BreakAll;
options.ja_zh = false;
let segmenter = LineBreakSegmenter::try_new_with_options(options).expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_str("Hello World").collect();
assert_eq!(&breakpoints, &[1, 2, 3, 4, 6, 7, 8, 9, 10, 11]);

Segment a Latin1 byte string:

use icu_segmenter::LineBreakSegmenter;

let segmenter = LineBreakSegmenter::try_new().expect("Data exists");

let breakpoints: Vec<usize> = segmenter.segment_latin1(b"Hello World").collect();
assert_eq!(&breakpoints, &[6, 11]);

Word Break

Segment a string with default options:

 use icu_segmenter::WordBreakIterator;

 let mut iter = WordBreakIterator::new("Hello World");
 let result: Vec<usize> = iter.collect();
 println!("{:?}", result);

Segment a Latin1 byte string:

use icu_segmenter::WordBreakIteratorLatin1;

let s = "Hello World";
let iter = WordBreakIteratorLatin1::new(s.as_bytes());
let result: Vec<usize> = iter.collect();
println!("{:?}", result);

Generating property table

Copy the following files to tools directory. Then run ./generate_properties.py in tools directory (requires Python 3.8+). Machine generated files are moved to src directory.

https://www.unicode.org/Public/UCD/latest/ucd/LineBreak.txt

https://www.unicode.org/Public/UCD/latest/ucd/EastAsianWidth.txt

Macros

break_iterator_impl

Base implementation of rule based segmenter

Structs

GraphemeBreakIterator

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

GraphemeBreakIteratorLatin1

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

GraphemeBreakIteratorUtf16

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

Latin1Char

LineBreakIterator

The struct implementing the Iterator trait over the line break opportunities of the given string. Please see the module-level documentation for its usages.

LineBreakOptions

Options to tailor line breaking behavior, such as for CSS.

LineBreakSegmenter

Utf16Char

WordBreakIterator

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

WordBreakIteratorLatin1

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

WordBreakIteratorUtf16

The struct implementing the Iterator trait over the segmenter break opportunities of the given string. Please see the module-level documentation for its usages.

Enums

LineBreakRule

An enum specifies the strictness of line-breaking rules. It can be passed as an argument when creating a line breaker.

WordBreakRule

An enum specifies the line break opportunities between letters. It can be passed as an argument when creating a line breaker.

Traits

LineBreakType

A trait allowing for LineBreakIterator to be generalized to multiple string iteration methods.