icu_segmenter

Module line

Structs

LineBreakIterator

Implements the Iterator trait over the line break opportunities of the given string.

LineBreakOptions

Options to tailor line-breaking behavior.

LineBreakTypeLatin1

LineBreakTypePotentiallyIllFormedUtf8

LineBreakTypeUtf8

LineBreakTypeUtf16

LineSegmenter

Supports loading line break data, and creating line break iterators for different string encodings.

Enums

LineBreakStrictness

An enum specifies the strictness of line-breaking rules. It can be passed as an argument when creating a line segmenter.

LineBreakWordOption

An enum specifies the line break opportunities between letters. It can be passed as an argument when creating a line segmenter.

StringBoundaryPosType ğŸ”’

Constants

AI ğŸ”’

AL ğŸ”’

B2 ğŸ”’

BA ğŸ”’

BB ğŸ”’

BK ğŸ”’

CB ğŸ”’

CJ ğŸ”’

CL ğŸ”’

CM ğŸ”’

CP ğŸ”’

CR ğŸ”’

EB ğŸ”’

EM ğŸ”’

EX ğŸ”’

GL ğŸ”’

H2 ğŸ”’

H3 ğŸ”’

HL ğŸ”’

HY ğŸ”’

ID ğŸ”’

ID_CN ğŸ”’

IN ğŸ”’

IS ğŸ”’

JL ğŸ”’

JT ğŸ”’

JV ğŸ”’

LF ğŸ”’

NL ğŸ”’

NS ğŸ”’

NU ğŸ”’

OP_EA ğŸ”’

OP_OP30 ğŸ”’

PO ğŸ”’

PO_EAW ğŸ”’

PR ğŸ”’

PR_EAW ğŸ”’

QU ğŸ”’

RI ğŸ”’

SA ğŸ”’

SG ğŸ”’

SP ğŸ”’

SY ğŸ”’

UNKNOWN ğŸ”’

WJ ğŸ”’

XX ğŸ”’

ZW ğŸ”’

ZWJ ğŸ”’

Traits

LineBreakType

A trait allowing for LineBreakIterator to be generalized to multiple string iteration methods.

Functions

handle_complex_language_utf8 ğŸ”’

handle_complex_language impl for UTF8 iterators

is_break_utf32_by_loose ğŸ”’

Type Aliases

LineBreakIteratorLatin1

Line break iterator for a Latin-1 (8-bit) string.

LineBreakIteratorPotentiallyIllFormedUtf8

Line break iterator for a potentially invalid UTF-8 string.

LineBreakIteratorUtf8

Line break iterator for an str (a UTF-8 string).

LineBreakIteratorUtf16

Line break iterator for a UTF-16 string.