icu_segmenter

Module options

Options structs and enums

Structs

LineBreakOptions

Options to tailor line-breaking behavior.

SentenceBreakInvariantOptions

Locale-independent options to tailor sentence breaking behavior

SentenceBreakOptions

Options to tailor sentence breaking behavior.

WordBreakInvariantOptions

Locale-independent options to tailor word breaking behavior

WordBreakOptions

Options to tailor word breaking behavior.

Enums

LineBreakStrictness

An enum specifies the strictness of line-breaking rules. It can be passed as an argument when creating a line segmenter.

LineBreakWordOption

An enum specifies the line break opportunities between letters. It can be passed as an argument when creating a line segmenter.

WordType

The word type tag that is returned by WordBreakIterator::word_type().