icu_segmenter

Module word

Structs

WordBreakIterator

Implements the Iterator trait over the word boundaries of the given string.

WordBreakTypePotentiallyIllFormedUtf8

WordBreakTypeUtf8

WordBreakTypeUtf16

WordSegmenter

Supports loading word break data, and creating word break iterators for different string encodings.

WordTypeULE

ULE type for WordType

Enums

WordType

The word type tag that is returned by WordBreakIterator::word_type().

Functions

handle_complex_language_utf8 ๐

handle_complex_language impl for UTF8 iterators

Type Aliases

WordBreakIteratorLatin1

Word break iterator for a Latin-1 (8-bit) string.

WordBreakIteratorPotentiallyIllFormedUtf8

Word break iterator for a potentially invalid UTF-8 string.

WordBreakIteratorUtf8

Word break iterator for an str (a UTF-8 string).

WordBreakIteratorUtf16

Word break iterator for a UTF-16 string.